+++ src/crypto/PopTokenGenerator.ts
@@ -4,11 +4,8 @@
  */
 
 import { ICrypto, SignedHttpRequestParameters } from "./ICrypto";
-import { AuthToken } from "../account/AuthToken";
-import { TokenClaims } from "../account/TokenClaims";
 import { TimeUtils } from "../utils/TimeUtils";
 import { UrlString } from "../url/UrlString";
-import { ClientAuthError } from "../error/ClientAuthError";
 
 /**
  * See eSTS docs for more info.
@@ -20,6 +17,12 @@ import { ClientAuthError } from "../error/ClientAuthError";
 type ReqCnf = {
     kid: string;
     xms_ksl: KeyLocation;
+};
+
+export type ReqCnfData = {
+    kid: string;
+    reqCnfString: string;
+    reqCnfHash: string;
 };
 
 enum KeyLocation {
@@ -35,11 +38,28 @@ export class PopTokenGenerator {
         this.cryptoUtils = cryptoUtils;
     }
 
-    async generateCnf(request: SignedHttpRequestParameters): Promise<string> {
+    /**
+     * Generates the req_cnf validated at the RP in the POP protocol for SHR parameters
+     * and returns an object containing the keyid, the full req_cnf string and the req_cnf string hash
+     * @param request
+     * @returns
+     */
+    async generateCnf(request: SignedHttpRequestParameters): Promise<ReqCnfData> {
         const reqCnf = await this.generateKid(request);
-        return this.cryptoUtils.base64Encode(JSON.stringify(reqCnf));
+        const reqCnfString: string = this.cryptoUtils.base64Encode(JSON.stringify(reqCnf));
+
+        return {
+            kid: reqCnf.kid,
+            reqCnfString, 
+            reqCnfHash: await this.cryptoUtils.hashString(reqCnfString) 
+        };
     }
 
+    /**
+     * Generates key_id for a SHR token request
+     * @param request
+     * @returns
+     */
     async generateKid(request: SignedHttpRequestParameters): Promise<ReqCnf> {
         const kidThumbprint = await this.cryptoUtils.getPublicKeyThumbprint(request);
 
@@ -49,22 +69,31 @@ export class PopTokenGenerator {
         };
     }
 
-    async signPopToken(accessToken: string, request: SignedHttpRequestParameters): Promise<string> {
-        const tokenClaims: TokenClaims | null = AuthToken.extractTokenClaims(accessToken, this.cryptoUtils);
-        if (!tokenClaims?.cnf?.kid) {
-            throw ClientAuthError.createTokenClaimsRequiredError();
-        }
-        
-        return this.signPayload(accessToken, tokenClaims.cnf.kid, request);
+    /**
+     * Signs the POP access_token with the local generated key-pair
+     * @param accessToken
+     * @param request
+     * @returns
+     */
+    async signPopToken(accessToken: string, keyId: string, request: SignedHttpRequestParameters): Promise<string> {
+        return this.signPayload(accessToken, keyId, request);
     }
 
-    async signPayload(payload: string, kid: string, request: SignedHttpRequestParameters, claims?: object): Promise<string> {
+    /**
+     * Utility function to generate the signed JWT for an access_token
+     * @param payload
+     * @param kid
+     * @param request
+     * @param claims
+     * @returns
+     */
+    async signPayload(payload: string, keyId: string, request: SignedHttpRequestParameters, claims?: object): Promise<string> {
+
         // Deconstruct request to extract SHR parameters
         const { resourceRequestMethod, resourceRequestUri, shrClaims, shrNonce } = request;
 
         const resourceUrlString = (resourceRequestUri) ? new UrlString(resourceRequestUri) : undefined;
         const resourceUrlComponents = resourceUrlString?.getUrlComponents();
-
         return await this.cryptoUtils.signJwt({
             at: payload,
             ts: TimeUtils.nowSeconds(),
@@ -75,6 +104,6 @@ export class PopTokenGenerator {
             q: (resourceUrlComponents?.QueryString) ? [[], resourceUrlComponents.QueryString] : undefined,
             client_claims: shrClaims || undefined,
             ...claims
-        }, kid);
+        }, keyId, request.correlationId);
     }
 }