+++ src/response/ResponseHandler.ts
@@ -272,8 +272,8 @@ export class ResponseHandler {
      */
     private generateAccountEntity(serverTokenResponse: ServerAuthorizationTokenResponse, idToken: AuthToken, authority: Authority, oboAssertion?: string, authCodePayload?: AuthorizationCodePayload): AccountEntity {
         const authorityType = authority.authorityType;
-        const cloudGraphHostName = authCodePayload ? authCodePayload.cloud_graph_host_name : "";
-        const msGraphhost = authCodePayload ? authCodePayload.msgraph_host : "";
+        const cloudGraphHostName = authCodePayload ? authCodePayload.cloud_graph_host_name : Constants.EMPTY_STRING;
+        const msGraphhost = authCodePayload ? authCodePayload.msgraph_host : Constants.EMPTY_STRING;
 
         // ADFS does not require client_info in the response
         if (authorityType === AuthorityType.Adfs) {
@@ -311,7 +311,7 @@ export class ResponseHandler {
         requestState?: RequestStateObject,
         code?: string
     ): Promise<AuthenticationResult> {
-        let accessToken: string = "";
+        let accessToken: string = Constants.EMPTY_STRING;
         let responseScopes: Array<string> = [];
         let expiresOn: Date | null = null;
         let extExpiresOn: Date | undefined;
@@ -320,7 +320,13 @@ export class ResponseHandler {
         if (cacheRecord.accessToken) {
             if (cacheRecord.accessToken.tokenType === AuthenticationScheme.POP) {
                 const popTokenGenerator: PopTokenGenerator = new PopTokenGenerator(cryptoObj);
-                accessToken = await popTokenGenerator.signPopToken(cacheRecord.accessToken.secret, request);
+                const { secret, keyId } = cacheRecord.accessToken;
+
+                if (!keyId) {
+                    throw ClientAuthError.createKeyIdMissingError();
+                }
+
+                accessToken = await popTokenGenerator.signPopToken(secret, keyId, request);
             } else {
                 accessToken = cacheRecord.accessToken.secret;
             }
@@ -353,7 +359,8 @@ export class ResponseHandler {
             state: requestState ? requestState.userRequestState : Constants.EMPTY_STRING,
             cloudGraphHostName: cacheRecord.account?.cloudGraphHostName || Constants.EMPTY_STRING,
             msGraphHost: cacheRecord.account?.msGraphHost || Constants.EMPTY_STRING,
-            code
+            code,
+            fromNativeBroker: false
         };
     }
 }